feat: Room.java, RoomRepository.java 삭제
ERD를 기준으로 만든 ChattingRoom.java, ChattingRoomRepository.java가 해당 클래스를 대체하도록 수정

ref: #28

+++ src/main/java/com/aliens/friendship/controller/ChatController.java
@@ -22,7 +22,7 @@ public class ChatController {
 
     @MessageMapping("/{roomId}") //여기로 전송되면 메서드 호출 -> WebSocketConfig prefixes 에서 적용한건 앞에 생략
     @SendTo("/room/{roomId}")   //구독하고 있는 장소로 메시지 전송 (목적지)  -> WebSocketConfig Broker 에서 적용한건 앞에 붙어줘야됨
-    public ChatMessage test(@DestinationVariable Long roomId, ChatMessage message) {
+    public ChatMessage test(@DestinationVariable Integer roomId, ChatMessage message) {
 
         //채팅 저장
         Chat chat = chatService.createChat(roomId, message.getSender(), message.getMessage());
